PMM-15228 Reduce stop wait time for supervisord processes
