--- managed/services/supervisord/devcontainer_test.go
@@ -1,75 +1,0 @@
-// Copyright (C) 2023 Percona LLC
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Affero General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU Affero General Public License for more details.
-//
-// You should have received a copy of the GNU Affero General Public License
-// along with this program. If not, see <https://www.gnu.org/licenses/>.
-
-package supervisord
-
-import (
-	"os"
-	"path/filepath"
-	"testing"
-	"time"
-
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-
-	"github.com/percona/pmm/managed/models"
-)
-
-// TODO move tests to other files and remove this one.
-func TestDevContainer(t *testing.T) {
-	t.Run("UpdateConfiguration", func(t *testing.T) {
-		// logrus.SetLevel(logrus.DebugLevel)
-		vmParams, err := models.NewVictoriaMetricsParams(models.BasePrometheusConfigPath, models.VMBaseURL)
-		require.NoError(t, err)
-
-		s := New("/etc/supervisord.d", &models.Params{VMParams: vmParams, PGParams: &models.PGParams{}, HAParams: &models.HAParams{}})
-		require.NotEmpty(t, s.supervisorctlPath)
-
-		// restore original files after test
-		originals := make(map[string][]byte)
-		matches, err := filepath.Glob("/etc/supervisord.d/*.ini")
-		require.NoError(t, err)
-		for _, m := range matches {
-			b, err := os.ReadFile(m)
-			require.NoError(t, err)
-			originals[m] = b
-		}
-		defer func() {
-			for name, b := range originals {
-				err = os.WriteFile(name, b, 0)
-				require.NoError(t, err)
-			}
-			// force update supervisor config
-			err = s.supervisorctl("update")
-			require.NoError(t, err)
-		}()
-
-		settings := &models.Settings{
-			DataRetention: 3600 * time.Hour,
-		}
-
-		b, err := s.marshalConfig(templates.Lookup("victoriametrics"), settings)
-		require.NoError(t, err)
-		changed, err := s.saveConfigAndReload("victoriametrics", b)
-		require.NoError(t, err)
-		assert.True(t, changed)
-		changed, err = s.saveConfigAndReload("victoriametrics", b)
-		require.NoError(t, err)
-		assert.False(t, changed)
-
-		err = s.UpdateConfiguration(settings)
-		require.NoError(t, err)
-	})
-}
