+++ managed/services/supervisord/supervisord_test.go
@@ -114,3 +114,49 @@ func TestConfigVictoriaMetricsEnvvars(t *testing.T) {
 		})
 	}
 }
+
+func TestDevContainer(t *testing.T) {
+	t.Run("UpdateConfiguration", func(t *testing.T) {
+		// logrus.SetLevel(logrus.DebugLevel)
+		vmParams, err := models.NewVictoriaMetricsParams(models.BasePrometheusConfigPath, models.VMBaseURL)
+		require.NoError(t, err)
+
+		s := New("/etc/supervisord.d", &models.Params{VMParams: vmParams, PGParams: &models.PGParams{}, HAParams: &models.HAParams{}})
+		require.NotEmpty(t, s.supervisorctlPath)
+
+		// restore original files after test
+		originals := make(map[string][]byte)
+		matches, err := filepath.Glob("/etc/supervisord.d/*.ini")
+		require.NoError(t, err)
+		for _, m := range matches {
+			b, err := os.ReadFile(m)
+			require.NoError(t, err)
+			originals[m] = b
+		}
+		defer func() {
+			for name, b := range originals {
+				err = os.WriteFile(name, b, 0)
+				require.NoError(t, err)
+			}
+			// force update supervisor config
+			err = s.supervisorctl("update")
+			require.NoError(t, err)
+		}()
+
+		settings := &models.Settings{
+			DataRetention: 3600 * time.Hour,
+		}
+
+		b, err := s.marshalConfig(templates.Lookup("victoriametrics"), settings)
+		require.NoError(t, err)
+		changed, err := s.saveConfigAndReload("victoriametrics", b)
+		require.NoError(t, err)
+		assert.True(t, changed)
+		changed, err = s.saveConfigAndReload("victoriametrics", b)
+		require.NoError(t, err)
+		assert.False(t, changed)
+
+		err = s.UpdateConfiguration(settings)
+		require.NoError(t, err)
+	})
+}
