+++ managed/testdata/supervisord.d/nomad-server.ini
@@ -8,7 +8,7 @@ autostart = true
 startretries = 10
 startsecs = 1
 stopsignal = INT
-stopwaitsecs = 300
+stopwaitsecs = 10
 stdout_logfile = /srv/logs/nomad-server.log
 stdout_logfile_maxbytes = 10MB
 stdout_logfile_backups = 3
