+++ managed/testdata/supervisord.d/vmproxy.ini
@@ -13,7 +13,7 @@ autostart = true
 startretries = 10
 startsecs = 1
 stopsignal = INT
-stopwaitsecs = 300
+stopwaitsecs = 10
 stdout_logfile = /srv/logs/vmproxy.log
 stdout_logfile_maxbytes = 10MB
 stdout_logfile_backups = 3
